+++ Core/include/Acts/TrackFitting/GlobalChiSquareFitter.hpp
@@ -667,8 +667,7 @@ class Gx2Fitter {
 
             const double sigma = Acts::computeMultipleScatteringTheta0(
                 interaction.slab, particle.absolutePdg(), particle.mass(),
-                static_cast<float>(
-                    parametersWithHypothesis->parameters()[eBoundQOverP]),
+                parametersWithHypothesis->parameters()[eBoundQOverP],
                 particle.absoluteCharge());
             ACTS_VERBOSE(
                 "        The Highland formula gives sigma = " << sigma);
